+++ src/game/items/container.ts
@@ -1,6 +1,6 @@
 import { createDynamicText } from "../../utils/dynamicDescription";
 import { Verb } from "../verbs/verb";
-import { Item, Builder as ItemBuilder, customiseVerbs } from "./item";
+import { Item, Builder as ItemBuilder, customiseVerbs, omitAliases } from "./item";
 
 export function newContainer(config: ContainerConfig & ItemConfig) {
   const {
@@ -42,6 +42,9 @@ export function newContainer(config: ContainerConfig & ItemConfig) {
 
   Object.entries(remainingConfig).forEach(([key, value]) => (container[key] = value));
   customiseVerbs(config.verbCustomisations, container);
+
+  // Remove unwanted aliases.
+  omitAliases(config.omitAliases, container);
 
   return container;
 }